Satellite Map Generation/requirements.txt
# pip install -r requirements.txt

numpy==1.22.2
Pillow==8.3.1
pyscreenshot==2.3
selenium==3.141.0

